Enable use of redis.conf in start_servers.sh
This change addresses issue #21 by allowing you to set an environment
variable (REDIS_CONF_PATH) to specify the path to a redis.conf file
that you want to use when you run start_servers.sh. If you don't set
the env var, then Redis will run as before using the built-in default
configuration.

This also updates the README file based on this change and adds
appendonly.aof to the .gitignore, which is needed if you turn on AOF
persistence in Redis as the README now recommends.