Fix sierra DB constraint issues with test fixtures
This changeset relaxes some of the constraints previously added to
Sierra (base) models based on inspecting the DB, which were apparently
in error. After extracting some data from our live Sierra DB to generate
test DB fixtures and trying to load those fixtures into the test DB
created based on the models, there were some, e.g., fields that needed
to allow for null values and the like. The models should now be closer
to the actual Sierra DB setup. As we add more test fixtures, we'll
likely find more of this in the future.