Change settings to use .env file for secrets
As a first step toward Dockerization, this changeset modifies how
secrets and local settings (DB names, hosts, passwords) are stored.
Docker-compose allows you to pull environment variables from a .env
file, and so it seems to make sense to use the same .env file format to
communicate secrets both to Django settings and to docker-compose. This
way we don't have to have separate settings files defining the exact
same settings.