Set up Docker dev/test databases
This set of changes adds the needed files to support running existing
tests via Docker--sets up both the sierra and default test databases
along with the dev default database, and includes a Dockerfile to use
for running the actual app in a container with the appropriate
connections. There's still a lot left to do to support fully
containerized dev/test environments: setting up testing and dev
instances of Redis, Solr, and Celery.