Fix multiple issues with Docker setup
There are a few minor teaks made with this changeset. The Dockerfile was
updated to include openjdk so that Solrmarc will run. The Solrmarc
dev_config.properties file was modified and a test_config.properties
file was added to provide the correct settings for Solrmarc running in
our containers (for dev and test, respectively). Finally, the commands
to run MariaDB in docker-compose.yml were updated to provide a longer
timeout value for connections. The default value was 600, which is only
10 minutes, and long Celery processes sometimes need more time than
that. I upped it to 7200 seconds, which is two hours.