Convert migrations for base app to data fixtures
Migrations for the base app (for the Sierra database) were taking five
minutes to run in Django 1.7.1, just because there are over 300 models.
To save time, I removed the migrations entirely so that migrating the
sierra DB will syncdb based on the model definitions, which is all we
really need. It's also MUCH faster and now runs in seconds. Fixtures
need to be installed manually with "loaddata," however.