Clean up Dockerfile and add dockerignore
One, this adds a .dockerignore file to ignore everything except what
we're COPYing over to the container, which is the requirements files.
Two, this greatly cleans up the Dockerfile. Commands are combined to
reduce the number of steps needed. Time-consuming apt-get installs now
appear first so that we can change the Dockerfile and rebuild it much
faster. Finally, added args for userid and groupid so that we can
(optionally) specify what user needs to run in the container by default,
to help with user / volume issues.