Specify user/group in every docker-compose service
This is to help with an issue I'm running into on Linux, especially
Travis, builds, where services running in containers are giving
permission denied errors when trying to write to data volumes mounted on
the local host. This helps ensure that containers will run under the
same user/group as the user running the docker-compose.sh script.