Add shell script for managing local Docker volumes
This gets rid of docker_run_migrations.sh and creates init-dockerdata.sh
in its place. The new script is much more robust than the old one,
allowing you to build/migrate and rebuild/remigrate data for particular
services. It helps ensure that the user that will be running
docker-compose.sh owns the data volumes, which in turn ensures the
correct uid/gid will be running inside the container for each service.
(Apparently this is not always the case when relying on Docker to create
data volumes for you.) It also takes care of making sure the Postgres
database (sierra-test-db) is created via the postgres user and then
switched over to the custom uid.