Fix sierra-test-db fixture loading on Linux
This fixes the shell command used in init-dockerdata.sh that finds and
then passes args to manage-test loaddata for fixtures. Was working on
Mac, but I noticed it was not working correctly on TravisCI Linux.
Removing the quotation marks around the find command and converting
newlines to spaces seems to have done the trick.