Add VirtualCompField and associated tests
This is a big commit, but it adds a new custom field type,
`base.fields.VirtualCompField`, a virtual composite field, that we can
use with Sierra models that are lacking a PK or even a truly unique
individual field. The field is virtual, so it isn't reflected in the
database itself.

This will allow us to fix some issues with a few of the models, where
we had defined a non-unique field as a PK just to satisfy Django's
requirements for models. Doing this led to inappropriately restricted
data in the Sierra test database, which means our test data wasn't
reflective of the live data.