Improve VirtualCompField and enable serialization
This is another large-ish commit that fixes some issues with the initial
VirtualCompField implementation. The biggest improvement is that field
values are no longer represented as composite strings, but tuples, which
makes more sense. Composite string values are still needed for some
database operations and for serialization, but the default separator
value has been changed to `|_|` to reduce the chances that you'd
legitimately encounter that value in your data. Support for
serialization has been added. (Normally we'd just tell it not to
serialize the field, but that doesn't work when the field is used as a
PK.)