Update requirements: psycopg2 and pytest
Two things.

First, we don't want to use the precompiled psycopg2
(psycopg2-binary) in a production environment, and, with the Docker
installation, we're installing the prerequisites for compiling psycopg2
psycopg2 from source *anyway*. So, no reason not to use the non-binary
version. However, to prevent it from installing the binary via wheel, we
have to specify the --no-binary option.
See http://initd.org/psycopg/docs/install.html#binary-install-from-pypi
for more information.

Second, I added an explicit requirement for pytest to requirements-tests,
for consistency.