Set IiiUserIiiRole.iii_user back to FK, from 1-1
Oops--IiiUserIiiRole is a `through` table on a M2M relationship and
should not have a one-to-one relationship with IiiUser. I caught this
by running the database tests against our live Sierra DB. I went back
through the models and confirmed that this was the only OneToOneField
that I missed changing back to a ForeignKey.