Add infrastructure and initial tests for `api` app
This implements the utils.test_helpers.solr_factories added in the last
several commits. We add conftest.py fixtures at various levels to
generate test data for testing the `api` app. Starts out with some
initial tests of DRF exceptions, which we know are broken with the
updated version of DRF.