Refactor paginated simpleview; catch SolrErrors
This is a slightly tangled changeset made with the ultimate goal of
catching SolrErrors at the filter level rather than letting them go
(uncaught) to the pagination level. I refactored what was the
simpleviews.SimpleGetMixin.paginate method and broke it up into several
smaller methods, which helps with readability and isolates the code that
paginates the queryset. Because querysets are evaluated lazily, filtered
querysets aren't sent to Solr until the pagination happens, but the
filter is responsible for validating data, so we want the filter to
catch Solr errors when a `search` filter is used. At the same time, we
don't want to send the same search to Solr twice. This refactor exposes
a `paginate_queryset` view method that the filter can use to trigger the
pagination to catch Solr errors, which also caches the page results
(temporarily) which the view itself uses then clears when rendering page
data.