Streamline tests requiring static data (in Solr)
The api filter tests are growing to the point that we need to store
parametrization data separately from the test function. This affords us
with two things: we can tie IDs directly to the test parameters to which
they apply, and we can more easily create different groupings of test
parameters. In this case, we have some filter behavior that we know to
be odd that I want to write explicit test cases for so that later we can
go back and work on making those cases behave more normally. So now we
have two groups of test parameters: NORMAL and STRANGE.

The other thing I've snuck into this commit is a new helper method in
utils.test_helpers.fixture_factories' SolrTestDataAssembler class, which
automates the common case of loading multiple test records containing
static data, such as what we have in our api filter tests.