Improve parsing of array values for in and range
Added the ability to use quotation marks and the backslash characters to
escape values in arrays for the `in` and `range` operators. Previously
arrays were comma-delimited lists of values but gave no way to use a
comma as part of an actual value. Surrounding a value in quotation marks
will force commas inside the quotation marks not to be counted as
delimiters. Backslash characters must now be used to escape literal
quotation marks. Backslashes may also be used to escape commas. Thus --
["a,b",c,d\,e,"\"f,g\""] translates to the Python list ['a,b', 'c',
'd,e', '"f,g"'].