Fix isnull and isnull exclusion filters
Continuing to add API tests for filters and fix issues. This fixes an
issue with the Solr test data generation when the `creator` field is
None so that we can test the `isnull` operator. It also fixes a strange
issue with how we were doing exclusion/negation. Previously, doing an
`exclude` filter would AND together the excluded criteria and wrap the
whole thing in -(). But, with an "-isnull=true" operator, that led to,
e.g., a Solr query like this: -(-creator:*) which for some reason
doesn't always work correctly. The solution was to loop through each
excluded query criterion and manually switch (add or remove) the hyphen.
So, "creator[isnull]=true" looks like this: -creator:* and
"creator[-isnull]=true" looks like this: creator:*.