Test standard api list/detail resource views
This commit adds some tests (along with helper functions) that ensure
the primary standard types of resources such as bibs, items, eresources,
etc. (that have list and detail views) work as expected. Links amongst
and between resources are tested, as well.

Additionally, this commit adds a `resource_name` attribute to API
detail view classes. This attribute appears already on list view classes
but was left off of detail views, probably as an oversight. It's helpful
for checking what resource is returned in test responses.