Fix/test callnumbermatches, firstitemperlocation
Specialized API resources `callnumbermatches` and `firstitemperlocation`
are on-the-fly resources that are more like different or
partially-filtered views of the `items` resource. They were created to
serve specific functions for the Inventory App. This commit fixes their
View objects so they align with changes made to api.views.simpleviews
earlier. It adds tests to ensure that they function as expected.