Refactor exporters and search indexes for testing
This is a big set of changes aimed to improve the testability of
exporters and (to a lesser extent) Haystack search indexes.

Search indexes include a few quality-of-life improvements, plus I've
added the ability to get and examine the Solr schema for a given index
via the Solr API. This lets me interrogate a Haystack index object for
details about the underlying Solr index--some of which may not be
reflected in the Haystack field setup. This makes it possible to
compare, e.g., an indexed test result set against the schema fields in
the index for stronger and more reliable result validation. Especially
for the bibdata and marc indexes, which aren't fully managed by
Haystack.

Exporters have been refactored slightly so that they are more DRY and
there are now standardized ways to store and access multiple Haystack
index objects (e.g. for BibsToSolr) and multiple child exporter
processes. The ToSolrExporter base class institutes an `index_settings`
class property, methods for generating Haystack index objects, and
methods for doing common index operations. All *ToSolr classes are now
subclasses of ToSolrExporter. A CompoundMixin mixin class has been added
for exporters that are "compound" in that they operate (in whole or
part) by calling other exporters. This mixin adds an `exporter_names`
class property for defining the children exporter classes plus methods
and properties for generating and accessing children exporter classes.
Finally, the AttachedRecordExporter base class is for the compound
exporter classes where there is a main "parent" record type and attached
"child" record types that are all updated when the main exporter runs.
This is for, e.g., BibsAndAttachedToSolr: bibs (BibsToSolr) is the main
/ parent exporter, and ItemsToSolr and HoldingUpdate are the children.

Exporter tests have been modified, as well. It is now possible for tests
to configure the export-process -> exporter-class mapping at the module
level, allowing, e.g., the `export` app exporter tests to test those
export processes and the `shelflist` app exporter tests to test export
processes that override those in the `export` app. Exporter tests now
utilize the extra Exporter and SearchIndex functionality to test exports
more completely and robustly without as much configuration (although the
tests themselves are more complicated). In the next commit I'll try to
tighten up the `test_exporters.test_exports_to_solr` code.

Finally, in preparation for writing some tests for the `shelflist` app
that will test outputting data to Redis, I snuck in a conftest.py hook
that makes sure to clear out the Redis appdata database after each test.
I was concerned at first this might slow down the tests, but it doesn't
seem to have any measurable affect.