Catch and handle index errors in full_prepare
For rare cases where there is an uncaught error while preparing an
object from the Sierra DB for indexing, this catches any such error in
the `full_prepare` method and raises an Exception that tells Haystack to
skip that record. Records skipped this way are logged as errors on the
exporter.