Limit AttachedRecordExporter.final_callback dels
The AttachedRecordExporter `export_records` and `delete_records` methods
now track the type of job running (e.g. export or delete) via `vals`.
Because deletions are only ever done on the `main_child` (not the
children representing attached records), we want to limit the
`final_callback` so that it only triggers the main_child final_callback,
not final_callbacks on all children.