Add Sierra test fixtures for deleted records
Export tests currently do not handle deletions exactly correctly;
test deletions are based on records that have not been deleted in Sierra,
which means the entire record (bib, item, whatever) is passed to the
deletion handler. In reality, only the RecordMetadata record for deleted
records is left in Sierra after a record is deleted; testing deletions
correctly means we have to pass only a stub RecordMetadata record from
deleted records to the deletion handler. The first step in correcting
this is to add sierra-test-db fixtures for deleted records (bibs, items,
and eresources).