Move shelflist manifest generation into index
This is the first step in refactoring how shelflist manifests are
generated. It makes a little more sense to have the SearchIndex object
track locations that have been added during index updates and then to
include methods for saving those manifests to Redis on the index object
rather than on the exporter object.

Making these changes, I realized that I need to rethink how shelflist
item manifest updates for deletions (delete_records) are done, so that
will be the next step.