Improve export prefetch_ and select_related props
I realized while working on the bl-alpha-solrmac branch that I could
make a few easy improvments to the prefetch_ and select_related
properties on most `basic_exporters` to help make a slight performance
improvement.

I also realized there's no reason to run HoldingsUpdate in production,
and it was running as part of BibsAndAttachedToSolr. This update
comments out HoldingsUpdate so it doesn't run. Should dramatically
improve BibsAndAttachedToSolr performance. As long as we're running a
full EResourceToSolr update every night, running HoldingsUpdate is
redundant.