Wrap tasks in a connection manager
Celery tasks run outside the normal Django request cycle and so don't
honor the CONN_MAX_AGE database setting by default. I occasionally run
into issues both with the Sierra and the default database, where
connections from tasks will be left open, get closed by the database,
and the next time a task runs it raises an error because the connection
it thinks is still open was closed unexpectedly. This change adds a
`manage_connections` decorator for tasks that makes sure connections are
closed correctly both before and after a job runs. It defers to
django.db.connections' `close_if_unusable_or_obsolete` which DOES use
the CONN_MAX_AGE setting to determine if a connection should be closed.
If so, it closes it gracefully so a new one will be opened for the next
request. This ___should___ make the weird OperationalError checks
unnecessary, but I'm putting that fix in a new commit just in case
they're still needed.