Refactor export.tasks
This is a major refactoring of the Celery tasks that run export jobs. It
improves performance and reliability, while fixing some subtle but
pernicious issues with how tasks were performed previously. The biggest
change is the implementation of the `JobPlan` object, which enables sets
of records to be divided into concrete batches and chunks, where the
plan for each chunk includes a list of PKs to export for that chunk. The
plan is cached in Redis separately from the Celery result broker, and
the plan is used to manage delegation of tasks to Celery, rather than
spawning gigantic sets of parallel tasks ("chords") and relying on
Celery to manage them.

This should fix longstanding issues with reliability of very large
export jobs. Plus, now, if a portion of an export job doesn't run, the
applicable parts of the plan are kept in Redis for troubleshooting. (The
plan is deleted if the entire job runs successfully.)

I have not yet deployed this to our staging/production environment, so
further bug fixes may be coming.