Fix annoying default kwarg logic in to_file
Previously, the S2MarcBatch.to_file method used logic in the method
signature, which is not good practice. I thought it was causing some odd
results involving filenames reusing old timestamps, but the problem
turned out to be something else. Still, I moved the logic into the
method itself, where it should be, just to rule out that as a possible
problem.