Deprecate AllToSolr exporter
`AllToSolr` is not something that we use in production and it is no
longer something that we would ever use in production; scheduling and
coordinating when to load various record types is more nuanced than a
simple "load everything at once" approach can accomplish. It is
sufficiently different than other exporters that it is no longer worth
the upkeep, so I am deprecating it. I'll plan to remove it entirely
sometime in the near future.