Continue aligning code/tests with merged changes
Cleans up a couple of lines in the shelflist ItemsToSolr exporter code
that I missed during the merge. Also makes exporter tests a little more
"unit"-y and slightly less reliant on fixtures to perform code that the
tests should be performing directly. This fits and works better given
the changes to tasks.