Handle empty record sets during exporter job init
Rarely, an exporter job will run in which no qualifying records are
fetched from Sierra. Recent changes to tasks (JobPlan implementation,
etc.) did not take this into account, leading to the error described in
issue #46. This adds a check during job initialization and logs an
appropriate message if no records were found, bypassing the code path
that raised the error.