Add IIILanguage rel to Bibs exp prefetch_related
I noticed that the BibsToSolr exporter is about 20% slower in the
update-dependencies-jt branch than the one in master, and after
doing a bit of profiling I traced it back to the MARC conversion
routines in sierra2marc. From there I traced it to the implementation of
the base.models.ModelWithAttachedName class, particularly with the
base.models.MaterialPropertyName model. The IIILanguage model is used to
filter the appropriate name in the appropriate language. I added this
table to the `prefetch_related` attr in the applicable exporter, which
made a small improvement.