Isolate Batch behavior from Attached behavior
Still working on implementing proper Batch exporters. This refactors the
CompoundMixin and AttachedRecordExporter classes and adds a
BatchExporter class. The behavior that was previously part of
CompoundMixin, which assumed derived classes needed to act on attached
records in non-standard ways, has been moved entirely into the
AttachedRecordExporter class. Now the default CompoundMixin behavior is
to assume that the parent exporter is simply a coordinator for the
children exporters, including in terms of `get_records`. `BatchExporter`
implements this in a straightforward way, which we can use for
refactoring the `AllMetadataToSolr` class.

In the future, we'll reimplement the `AttachedRecordExporter` class and
subclasses as proper `BatchExporter`s; we just need to create an
appropriate filter in `base.managers` that will filter to the attached
records in a certain set. This set of changes helps us start preparing
for that.