Add "set_raw_params" method to Solr Queryset
The Solr Queryset object is a useful abstraction for querying and
filtering Solr data; however, it's also useful to be able to add raw
parameters to send to Solr (e.g. for faceting) in conjunction with a
filtered Queryset, which there was no good, simple way to do. This
rectifies that by adding a `set_raw_params` method. The `search` method
has also been updated to utilize the new method.