Add test for shelflistitem user fields
Running a shelflist.exporters.ItemsToSolr `export_records` job on a set
of records from Sierra that are already in Solr and have user-provided
fields (inventory_date, shelf_status, etc.) populated in Solr should
leave those fields alone but update the rest of the record. This adds a
test to ensure that is the case.