Decouple api fixtures from assemble_test_records
In order to use the `assemble_test_records` fixture in non-api-app
tests, I had to remove the `api_solr_env` and `basic_solr_assembler`
fixtures, and then create an api-specific `assemble_api_test_records`
fixture in the api-app tests that did use those fixtures. This fixes a
bug in the shelflist-api tests (not yet committed) where too much test
data was being created because I was inadvertently calling the
`api_solr_env` fixture without wanting to or realizing it.