Add first round of tests for shelflist app api
This adds a set of basic tests for the api resources that the
`shelflist` app adds or modifies: primarily `shelflistitems`, but also
items and locations. Necessary fixtures are added to the shelflist
conftest file. A tiny change to the shelflist views to facilitate
testing. More specific tests that test shelflist-specific features (like
updating/saving data to the API) are coming next.