Move Solrmarc conversion into custom Solr backend
As we're working on additional indexes fed by Solrmarc (for Blacklight),
it started making more sense to move the Solrmarc conversion process
into a lower-level class so that we can have more uniformity at the
index and exporter levels. This makes that change. It also means we can
remove some cruft from the solr/solrmarc project directory.