Remove cruft from solrmarc directory
Following from the last commit, this removes unnecessary properties
files from the solr/solrmarc project directory. It also updates the
remaining `config.properties` files to remove the illusion that the
`index.properties` file is still defined there. Now it is defined more
granularly in the haystack index class or the Solr backend.