Give CompoundMixin the ability to formulate rels
Previously I had moved some of the methods for formulating
combined `select_` and `prefetch_related` lists for multiple children
from `CompoundMixin` down into the `AttachedRecordExporter`, because I
thought the latter class would be the only place where that was really
useful. In working on blacklight exporters, I discovered this is not the
case, so I've moved that functionality back up into `CompoundMixin`, but
a little more generalized. The functionality that IS specific to
`AttachedRecordExporter` has been moved into the associated `Child`
class, for determining relationship strings for attached children.