Allow SolrmarcIndexBackend to ignore empty recsets
Previously if an index update operation was called for an empty record
set, the SolrmarcIndexBackend would still generate a MARC file (empty)
and attempt to run the Solrmarc Java process. This would lead to a weird
"divide by zero" error simply due to the empty file. This just adds
checks to make sure the file isn't empty before the Solrmarc process is
called.