Move importapiusers mgt command to api app
I had originally put the `importapiusers` management command in the
`sierra` app in order to keep it with the other commands, but this one
actually needs to be run in production, and the `sierra` app is not
installed there. It makes more sense to keep it with the `api` app, so
I've moved it there.