Fix shelflistitem.views.FirstItemPerLocationList
This corrects a small oversight in which I neglected to realize that the
`shelflist` app implemented a specialized/extended version of
`api.views.FirstItemPerLocationList`. When refactoring
`api.simpleviews`, I broke the `SimpleGetMixin` `paginate` method into
several smaller methods. I updated the
`api.views.FirstItemPerLocationList` view to use the new method, but I
didn't update the `shelflist.views.FirstItemPerLocationList`, and I
didn't add tests to test it. This corrects that mistake--it addes the
appropriate tests and fixes the issue by simply renaming that method on
the `shelflist` version of the view.