Add ability to customize task management
This refactors Celery tasks slightly to give exporters the ability to
customize how tasks are managed. Before, we assumed a context where all
exporters were loading records via Django ORM models, and so the
export.tasks.JobPlan class assumed it would be working with explicit
lists of PKs to break export jobs into batches of discrete tasks.
However, with bl-suggest, we want to write exporters that load from one
Solr instance into another. With a solr queryset, it's much more
performant to divide jobs into batches via start/end keys for each
batch.

Ultimately, I removed most of the weird private functions that we were
using to do the initial creation of PK lists and created a
`RecordSetBundler` class to take over that work. Essentially you just
create a subclass that defines how to convert a recordset or queryset
into a `bundle` payload (e.g. a list of PKs or a start ID, end ID, and
length) and back into a queryset slice for processing. You can now
define the bundler that should be used for an Exporter using the
`bundler` property.

Lastly, to simplify bundling of multi-recordset batches, I changed how
those are handled. The name of the queryset is no longer zipped into the
list of PKs, but rather they're handled as totally separate facets of a
given job, which makes much more sense.