Add initialize method for Exporters
This adds an initialize method to Exporters that can be run before
starting an export job (e.g. via Celery). It's equivalent to
`final_callback` but runs before the job starts.

For `ToSolrExporter` classes, the initialize method issues a delete *:*
command (without committing) if the export_type is 'full_export'. This
ensures a clean re-index.