Enable dynamic fields in Solr test profiles
This is a set of changes to let us define fields in Solr test profiles
that are represented in a Solr schema as dynamicFields. To support this
change, I had to remove the concept of using inclusive/exclusive lists
of fields when defining the `user_fields` in a
`solr_factories.SolrProfile` object. Now a list of `user_fields` is
ALWAYS considered inclusive. This is how all of our Solr test profiles
are already defined so I only had to change a few tests along with test
fixtures that used the `inclusive` parameter.