Tweak bl-suggest packing strategy
Still trying to iron out issues with memory-usage spikes caused by large
documents / updates. This slightly changes how bundles are created while
packing a job, so that packages are created based on counts of records
that have a particular value rather than the number of values
themselves. This should help even out some of the spikes.