Begin isolating pipeline for custom field data
In the sierra2marc code, we need to separate the action of generating
the standard MARC record (from Sierra data) and modifying it to shoehorn
our custom data in so that SolrMarc loads our custom data into Solr.
This commit sets that up (including tests for our customizations).

Later, when we move away from SolrMarc, we can hopefully take that
custom pipeline and use it as-is without too much modification.