Improve and test publication_display info
This adds some extensive testing of the publication display information
that BlacklightASMPipeline.get_pub_info generates. In the process, it
fixes a number of small problems.

Additionally I've added a `breaktags` kwarg option to the
`group_subfields` function so that you can now define one or more
subfields that will always trigger a new grouping. (260c, for example.)