Try to fix issues with copyright dates
In terms of publication information, copyright dates aren't handled very
well, partly because of how they're treated in the 260 vs the 264. They
also have a variety of allowable phrases that can stand in for the
copyright symbol. (And even different copyright symbols, depending on
whether it's a sound recording copyright or not.) And they may, rarely,
have additional text/phrasing that goes along with them (such as
"copyright 1964 by Such and Such"). Previously the rest of the
information was being stripped.

This is an attempt at both lessening the strictness of how copyright
dates were parsed (so that additional phrases are included) and also
ensuring that copyright symbols get converted to the correct symbol, as
appropriate.