Add module for creating custom local rulesets
This adds `base.ruleset`, which contains classes for creating rulesets
that let you reason about objects based on obj attributes. It's intended
these be used for modeling Sierra business rules that are local to a
given institution--such as ones related to various Sierra models but NOT
in `base.models` because they are not part of default Sierra.