Fix issue with regex causing certain recs to hang
One of the regexes that attempted to strip punctuation and whitespace
from the ends of the string was causing Python to hang while compiling
the regex when the string had ~14-15 consecutive spaces. Which
apparently some of our publication statements have. This improves that
regex to fix the problem, and adds a test case for it.