Ensure text-type query analyzers work correctly
This adds an apparently missing step to the BLAS `text` type query
analyzers in the schema. A `WordDelimiterFactory` was in place in the
index analyzers but missing from the query analyzers, which led to some
slightly strange and unexpected behavior where e.g. punctuation in a
search query would eliminate hits from records that didn't include the
punctuation. I think my intention was that we want to boost hits where
punctuation in the query exactly matches punctuation in the indexed
document, but we don't want to eliminate those hits. Adding the same
step to the query analyzer should still accomplish the former, since
it's set to keep the old and new tokens. `facet-search` is analyzed as
`facet-search` `facet` `search` and `facetsearch`. So a query
containing any of those would match, but `facet-search` would be
analyzed the same way and be a better match.