Make model `get_call_number` method more flexible
This fixes the base.models.ItemRecord and BibRecord `get_call_number`
method so that the `spec` variables used to define what varfields are
call numbers are mutually exclusive--a wildcard `c` tag rule following
specific `c` tag rules will not lead to pulling the same fields twice;
instead, they will be excluded. (It just changes the looping structure
and adds a break so that once a varfield matches one of the rules, it
goes to the next field.)

The ultimate goal was to allow logic like: a c-tagged 050, 055, or 090
field is an LC call number; any other c-tagged field is a Local call
number. (Which is how c-tagged fields on Item Records should work.)

I also added tests for the `get_call_number` methods.