Fix resource and media_type categorization
This fixes two minor issues with the resource and media_type
categorization.

1. Certain formats with spaces in them (e.g. 'Xbox 360') were coming
through with underscores in place of spaces. I decided against
underscores earlier but neglected to update everything.

2. Previously "software" types (bcode2 type 'm') defaulted to
a non-game computer program type. This changes that so that they
default to computer documents (digital files) rather than programs; if
008/26 is 'b' or 'i' it still comes through as non-game programs.