Deduplicate author/contributor names
Names that may be duplicated in 1XX/7XX fields for legitimate reasons
are now deduplicated when generating author/contributor values. A name
in a 1XX field that becomes the `author` that is repeated in a 7XX
field as part of a name/title heading (i.e. for related titles by the
same author) will ONLY display as the author and not as a contributor.
Similarly, any names in 7XXs that are repeated (i.e. for multiple
related titles) no longer generate duplicate contributor names.

Fixes discover issue 145.