Update period protection for longer numbers
Periods following numbers are ambiguous and difficult to say for sure
whether they are incidental or not. This adds a check for numbers with
4+ digits, where a lower-case letter following a number with a period
means it's more likely to be incidental. Previously I only included
1-3 digit numbers in the check for a period after an ordinal number
because a 4-digit number is more likely to be a date and not an
ordinal number, which is (maybe) slightly more likely to end with a
structural period. But, the Early English Books Online collection has
titles including years with periods after them, which often seem to
function the same way as periods following ordinal numbers.