Add parser for 245s (transcribed titles)
This adds a TranscribedTitleParser for parsing out 245s into meaningful
structures, including one or more transcribed titles, optional parallel
titles, and statements of responsibility. Each title is parsed into
hierarchical parts (if applicable) based on, e.g., subfields n or p.
Statements of responsibility are associated with the title they are
most likely to correspond with, but they are not duplicated (i.e. in
case of multiple titles or titles and parallel titles that share one
SOR.

Light formatting is applied when needed, particularly when values from
multiple subfields are joined together and appear in the same part of a
title but don't have (e.g.) punctuation between them, such as $p
following a $n but lacking a separating period or comma. This happens
most often with the archival-related subfields, fgks. $g is a special
case: if a $g (bulk dates) follows $f (dates) without any separating
punctuation, a display constant, e.g. `(bulk YYYY-YYYY)`, is added to
the $g data. This converts something like:

"Smith family papers 1953-2001 1953-1957"

to:

"Smith family papers, 1953-2001 (bulk 1953-1957)"

ISBD punctuation is used if present but not strictly required; a best-
guess attempt at parsing is done based on subfields alone if ISBD
punctuation is missing. Things like parallel titles of course can only
be found if the appropriate ISBD punctuation is included (' = '). Most
ISBD punctuation used during parsing gets stripped out. Periods are
used if it can be determined that they are not "incidental" (i.e. part
of an abbreviation, ordinal number, roman numeral, etc.), but this is
mainly the case when parsing $c (everything after the first SOR) and
when periods occur immediately preceding certain subfields (like $n or
$p) to indicate that the next subfield begins a new hierarchical part.

MARC rules that the $b must follow the first ISBD punctuation mark and
$c must start with the first SOR are not strictly enforced: if $a
contains an ' = ' or ' / ' or ' ; ', then parallel titles, SORs, and
additional titles are parsed out as needed. Periods are ignored in the
middle of $a and $b; we have too many titles with full sentences in
these subfields to assume that periods in these fields are intended to
denote hierarchal parts. If I find these settings too lax or strict
when working with real data, I'll tweak them.